fix to HIKe VM Core for supporting PER-CPU Shared Memory
caveat: every eBPF/HIKe Program SHOULD BE the single HIKe Program in a
single compilation unit (.bpf.o). Otherwise, if many different HIKe Progs
are within the same file, the compiler/optimizer can rearrange the code
in a way that it may not satisfy the verifier. As a consequence, the
verifier could reject the programs.

Signed-off-by: Anonymous Author <hike.eclat@gmail.com>